pyproject.toml
[project]
name = "face-diseases"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.5.5",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "tensorboard>=2.20.0",
    "termcolor>=3.3.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
